Latest changes for 1.3.12

+++ fresh-forms-for-gravity.php
@@ -3,7 +3,7 @@
  * Plugin Name: Fresh Forms for Gravity
  * Description: Prevent posts and pages with a Gravity Forms shortcode or Gutenberg block from being cached.
  * Author: Samuel Aguilera
- * Version: 1.3.5
+ * Version: 1.3.12
  * Author URI: https://www.samuelaguilera.com
  * Text Domain: fresh-forms-for-gravity
  * Domain Path: /languages
@@ -26,10 +26,90 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-define( 'FRESH_FORMS_FOR_GRAVITY_VERSION', '1.3.5' );
+define( 'FRESH_FORMS_FOR_GRAVITY_VERSION', '1.3.12' );
+
+// Scripts handlers for plugins using them for exclusion filters (e.g. SG Optimizer or Hummingbird).
+$fffg_js_handlers = array(
+	'jquery', // Yeah, not a GF script but many of them (and themes, and etc...) have it as dependency.
+	'jquery-core',
+	'gform_gravityforms',
+	'gform_conditional_logic',
+	'gform_datepicker_init',
+	'plupload-all', // Multi-upload fields.
+	'gform_json',
+	'gform_textarea_counter',
+	'gform_masked_input',
+	'gform_chosen',
+	'gform_placeholder',
+	'gforms_zxcvbn', // Password strength.
+	'gf_partial_entries', // Partial Entries Add-On.
+	'stripe.js', // Stripe Add-On.
+	'stripe_v3',
+	'gforms_stripe_frontend',
+	'gform_coupon_script', // Coupons Add-On.
+	'gforms_ppcp_frontend', // PPCP Add-On.
+	'gform_paypal_sdk', // Dependency for PPCP.
+	'wp-a11y', // Dependency for PPCP. This and the following three lines fixed issues with a PPCP form.
+	'wp-dom-ready', // Dependency for wp-a11y.
+	'wp-polyfill', // Dependency for wp-a11y.
+	'wp-i18n', // Dependency for wp-a11y.
+	'gforms_square_frontend', // Square Add-On.
+	'gform_mollie_components', // Mollie Add-On.
+	'gform_chained_selects', // Chained Selects Add-On.
+	'gsurvey_js', // Survey Add-On.
+	'gpoll_js', // Polls Add-On.
+	'gaddon_token', // Credit Card Token.
+	'gform_signature_frontend', // Signature Add-on.
+	'super_signature_script',
+	'super_signature_base64',
+	'gform_signature_delete_signature',
+	'gform_recaptcha', // reCAPTCHA.
+);
+
+// Scripts partial matches for plugins using them for exclusion filters (e.g. WP-Optimize or WP Rocket).
+$fffg_js_partial = array(
+	'gravityforms', // This is enough to match any script having gravityforms as part of the URL.
+	'jquery.min.js',
+	'plupload.min.js',
+	'a11y.min.js',
+	'wp-polyfill.min.js',
+	'dom-ready.min.js',
+	'i18n.min.js',
+	'zxcvbn.min.js', // Password strength.
+	'recaptcha',
+);
+
+// Inline scripts string matches for plugins using them for exclusion filters (e.g. SG Optimizer or WP Rocket).
+$fffg_js_inline_partial = array(
+	'gformRedirect',
+	'var gf_global',
+	'gformInitSpinner',
+	'var gf_partial_entries',
+	'(function(d,s,i,r)', // HubSpot Tracking Script.
+	'gform.addAction',
+	'gform_post_render',
+	'var gforms_ppcp_frontend_strings', // PPCP Add-On.
+	'gform_page_loaded', // Multi-page Ajax forms.
+	'var stripe', // Stripe Checkout.
+	'gform_gravityforms-js-extra',
+);
+
+// Domains for external JS for exclusion filters (e.g. SG Optimizer or WP Rocket).
+$fffg_js_external_domain = array(
+	'2checkout.com',
+	'agilecrm.com',
+	'dropbox.com',
+	'js.hs-analytics.net', // HubSpot Analytics Code.
+	'mollie.com',
+	'paypal.com',
+	'js.squareup.com',
+	'js.squareupsandbox.com',
+	'js.stripe.com',
+);
 
 add_action( 'gform_loaded', array( 'Fresh_Forms_For_Gravity_Bootstrap', 'load' ), 5 );
 register_activation_hook( __FILE__, 'fffg_purge_all_cache' );
+add_action( 'wp_loaded', 'fffg_actions_after_update' );
 
 /**
  * Class Fresh_Forms_For_Gravity_Bootstrap
@@ -81,6 +161,7 @@ function fresh_forms_for_gravity() {
  * Purge everything for a fresh start... This is required in order to allow donotcache_and_headers() to run before caching a page.
  */
 function fffg_purge_all_cache() {
+	global $file_prefix, $kinsta_cache, $epc; // Third-party variables that we may need.
 
 	// WP Engine.
 	if ( class_exists( 'WpeCommon' ) ) {
@@ -106,8 +187,6 @@ function fffg_purge_all_cache() {
 
 	// WP Super Cache.
 	if ( function_exists( 'wp_cache_clean_cache' ) ) {
-		global $file_prefix;
-
 		empty( $file_prefix ) ? $file_prefix = 'wp-cache-' : $file_prefix;
 		wp_cache_clean_cache( $file_prefix, true );
 	}
@@ -162,11 +241,37 @@ function fffg_purge_all_cache() {
 	}
 
 	// Kinsta Cache.
-	if ( class_exists( 'Kinsta\Cache' ) ) {
+	if ( class_exists( 'Kinsta\Cache' ) && is_object( $kinsta_cache ) ) {
 		// $kinsta_cache object already created by Kinsta cache.php file.
-		global $kinsta_cache;
 		$kinsta_cache->kinsta_cache_purge->purge_complete_full_page_cache();
 	}
 
+	// Endurance Page Cache.
+	if ( class_exists( 'Endurance_Page_Cache' ) && is_object( $epc ) ) {
+		// $epc object already created by endurance-page-cache.php file.
+		$epc->purge_all();
+	}
 }
 
+/**
+ * Actions to perform after an update.
+ */
+function fffg_actions_after_update() {
+	// Get stored version if any.
+	$current_version = get_option( 'fffg_version' );
+
+	// Return without actions if stored version match with defined version.
+	if ( version_compare( $current_version, FRESH_FORMS_FOR_GRAVITY_VERSION, '==' ) ) {
+		return;
+	}
+
+	/**
+	 * Try to flush cache to make sure changes made to new version are applied.
+	 * Known issue: WP-Optimize classes are not loaded at this point.
+	 */
+	fffg_purge_all_cache();
+
+	// Update stored version.
+	update_option( 'fffg_version', FRESH_FORMS_FOR_GRAVITY_VERSION );
+
+}
